script/tiled_bitmap_maker/requirements.txt
Pillow==7.2.0
absl-py==0.10.0
